utf8 font and character decoding working

+++ libraries/epd2in7b/fonts.h
@@ -68,8 +68,8 @@ typedef struct
 typedef struct
 {
 	const uint8_t 			heightPages;	// height, in pages (8 pixels), of the font's characters
-	const uint8_t 			startChar;		// the first character in the font (e.g. in charInfo and data)
-	const uint8_t 			endChar;		// the last character in the font
+	const uint32_t 			startChar;		// the first character in the font (e.g. in charInfo and data)
+	const uint32_t 			endChar;		// the last character in the font
 	const uint8_t			spacePixels;	// number of pixels that a space character takes up
 	const FONT_CHAR_INFO*	charInfo;		// pointer to array of char information
 	const uint8_t*			data;			// pointer to generated array of character visual representation
